requirements.txt
tensorflow==2.20.0
gradio==6.6.0
numpy
pillow